# docs/openapi.yaml
openapi: 3.0.1
info:
  title: Bank Card Management API
  description: REST API for managing bank cards
  version: "1.0"
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /api/cards:
    get:
      tags: [Cards]
      summary: Get my cards with optional status filter and pagination
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/CardStatus'
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
        - name: sort
          in: query
          required: false
          schema:
            type: string
            example: id,asc
      responses:
        '200':
          description: Paginated list of user's cards
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageCardResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cards/transfer:
    post:
      tags: [Cards]
      summary: Transfer funds between my cards
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
            example:
              fromCardId: 1
              toCardId: 2
              amount: "500.00"
      responses:
        '200':
          description: Transfer completed successfully
        '400':
          description: Insufficient funds or card not active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Card not found (or does not belong to user)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Concurrent modification conflict — retry the request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Source and target card are the same
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/cards/{id}/block:
    patch:
      tags: [Cards]
      summary: Request to block my card
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Block request submitted
        '404':
          description: Card not found (or does not belong to user)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/admin/cards/{id}/block:
    patch:
      tags: [Admin]
      summary: Block a card
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Card blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardResponse'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TransferRequest:
      type: object
      required: [fromCardId, toCardId, amount]
      properties:
        fromCardId:
          type: integer
          format: int64
          example: 1
        toCardId:
          type: integer
          format: int64
          example: 2
        amount:
          type: number
          minimum: 0.01
          example: 500.00
    CardResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        maskedCardNumber:
          type: string
          example: "**** **** **** 3456"
        ownerEmail:
          type: string
          example: user@bank.com
        expiryDate:
          type: string
          format: date
          example: "2028-12-31"
        status:
          $ref: '#/components/schemas/CardStatus'
        balance:
          type: number
          example: 1500.00
...
    CardStatus:
      type: string
      enum: [ACTIVE, BLOCKED, EXPIRED]
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          example: "2026-03-02T12:00:00"
        status:
          type: integer
          example: 404
        message:
          type: string
          example: "Карта не найдена: 99"
...
    PageCardResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CardResponse'
        totalElements:
          type: integer
          example: 5
        totalPages:
          type: integer
          example: 1
        size:
          type: integer
          example: 10
        number:
          type: integer
          example: 0
...